# Improve Frontend::apply_patch performance