… automerge-frontend/src/state_tree/diff_application_result.rs
… automerge-frontend/src/state_tree/diffable_sequence.rs
… automerge-frontend/src/state_tree/focus.rs
… automerge-frontend/src/state_tree/mod.rs
… automerge-frontend/src/state_tree/multivalue.rs
… automerge-frontend/src/state_tree/resolved_path.rs
… automerge-frontend/src/state_tree/state_tree_change.rs
… automerge/Cargo.toml
… automerge/benches/frontend_patch.rs