I've been testing this with a patch built from 159 changes. The `apply_patch` operation was taking ~720ms to finish (which was getting rather noticeable) but with this PR it now takes just ~52ms.

I believe my changeset was rather object heavy (rather than just a text value) which may have been more of an issue for the logic. I have added some benchmarks for the `apply_patch` operation to cover some of these cases.

From `perf` I found that lots of time was spent doing `malloc` and `free`. Then DHAT helped give a picture of where lots of the allocations were coming from and it turned out to be lots of cloning. So I've been through and tried to reduce the clones by either passing by value (rather than ref) or re-using existing datastructures (in the case of `current_objects`).

I think there is still room for improvement but wanted to do it bit by bit so it should be easier to check correctness.

Here are the benchmark results compared to main (https://github.com/automerge/automerge-rs/commit/ae56f0d12f1ed599720219b6246c1659d839b486) with the extra benchmarks.

```
B1.1 Append N characters
                        time:   [823.11 ms 825.63 ms 828.38 ms]
                        change: [-4.6630% -4.2938% -3.8849%] (p = 0.00 < 0.05)
                        Performance has improved.

B1.2 Append string of length N
                        time:   [61.911 ms 62.138 ms 62.317 ms]
                        change: [-19.129% -18.232% -17.388%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) low severe

B1.3 20√N clients concurrently set number in Map
                        time:   [436.28 us 440.28 us 444.61 us]
                        change: [-17.639% -16.373% -15.127%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

Frontend::apply_patch, small patch
                        time:   [1.5974 ms 1.6159 ms 1.6243 ms]
                        change: [-39.633% -37.435% -35.191%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  2 (20.00%) low mild

Frontend::apply_patch, small patch, many changes
                        time:   [1.5724 ms 1.5954 ms 1.6098 ms]
                        change: [-39.902% -37.844% -36.178%] (p = 0.00 < 0.05)
                        Performance has improved.

Frontend::apply_patch, nested maps
                        time:   [109.52 ms 110.49 ms 111.75 ms]
                        change: [-83.156% -82.963% -82.744%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) high mild
  1 (10.00%) high severe

Frontend::change insert long string
                        time:   [2.7435 ms 2.7906 ms 2.8168 ms]
                        change: [-17.015% -15.666% -14.331%] (p = 0.00 < 0.05)
                        Performance has improved.

StateTreeValue::apply_diff sequential text inserts across multiple patches
                        time:   [220.76 ms 221.50 ms 222.19 ms]
                        change: [-7.3750% -6.9084% -6.4340%] (p = 0.00 < 0.05)
                        Performance has improved.

StateTreeValue::apply_diff sequential text inserts in a single patch
                        time:   [9.5635 ms 9.6017 ms 9.6321 ms]
                        change: [-36.976% -35.693% -34.027%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) low mild
  1 (10.00%) high severe
```