@jeffa5 2021-04-06
I think we could hold off merging this until https://github.com/automerge/automerge-rs/pull/88 is done as it will likely change some of this stuff and best to just focus on the performance of the updated version.